requirements.txt
binance-connector==1.11.0
python-telegram-bot